Fix: flickering when load report
